pyproject.toml
[project]
name = "podman-pytest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "assertpy>=1.1",
    "docxtpl>=0.20.1",
    "gym>=0.26.2",
    "holidays>=0.73",
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "ollama>=0.5.1",
    "openapi-cli-generator>=0.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "ping3>=4.0.8",
    "pip>=25.1.1",
    "pytest>=8.3.5",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "scapy>=2.6.1",
    "scikit-learn>=1.6.1",
    "spacy==3.8.7",
    "streamlit>=1.45.1",
    "streamlit-autorefresh>=1.0.1",
    "umap-learn>=0.5.7",
]
